requirements.txt
requests==2.31.0
pandas==2.0.3
python-dotenv==1.0.0
numpy==1.24.3
plotly==5.13.1
schedule==1.2.1 